# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Việt phục Remix API",
    "description": "Backend REST API phục vụ nền tảng phối đồ Việt phục Remix (VietStylist)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/outfits": {
      "get": {
        "tags": [
          "Outfits & Versions"
        ],
        "summary": "List User Outfits",
        "description": "Lấy danh sách các bộ phối của người dùng đã đăng nhập.",
        "operationId": "list_user_outfits_api_outfits_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutfitResponse"
                  },
                  "title": "Response List User Outfits Api Outfits Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Outfits & Versions"
        ],
        "summary": "Create Outfit",
        "description": "Lưu bộ phối mới (hỗ trợ cả khách và người dùng đã đăng nhập).",
        "operationId": "create_outfit_api_outfits_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOutfitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutfitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/outfits/{outfit_id}": {
      "get": {
        "tags": [
          "Outfits & Versions"
        ],
        "summary": "Get Outfit",
        "description": "Xem thông tin chi tiết và snapshot hiện hành của bộ phối.",
        "operationId": "get_outfit_api_outfits__outfit_id__get",
        "parameters": [
          {
            "name": "outfit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Outfit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutfitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Outfits & Versions"
        ],
        "summary": "Update Outfit",
        "description": "Cập nhật bộ phối kèm kiểm tra xung đột revision phiên bản (Optimistic Concurrency).",
        "operationId": "update_outfit_api_outfits__outfit_id__put",
        "parameters": [
          {
            "name": "outfit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Outfit Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOutfitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutfitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Outfits & Versions"
        ],
        "summary": "Delete Outfit",
        "description": "Xóa bộ phối (Soft delete).",
        "operationId": "delete_outfit_api_outfits__outfit_id__delete",
        "parameters": [
          {
            "name": "outfit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Outfit Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/outfits/compare": {
      "post": {
        "tags": [
          "Outfits & Versions"
        ],
        "summary": "Compare Outfits",
        "description": "So sánh độc lập hai phương án A/B và hiển thị các điểm khác biệt (F08).",
        "operationId": "compare_outfits_api_outfits_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media/uploads": {
      "post": {
        "tags": [
          "Cloudflare R2 Media Management"
        ],
        "summary": "Request Upload Url",
        "description": "Tạo phiên tải lên file ảnh/video và cấp URL upload có hạn (Direct to R2) (F14).",
        "operationId": "request_upload_url_api_media_uploads_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUploadUrlInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/{media_id}/complete": {
      "post": {
        "tags": [
          "Cloudflare R2 Media Management"
        ],
        "summary": "Complete Upload",
        "description": "Xác nhận hoàn tất upload, kiểm tra HEAD và metadata trên storage trước khi chuyển trạng thái 'ready'.",
        "operationId": "complete_upload_api_media__media_id__complete_post",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAssetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/{media_id}/access": {
      "get": {
        "tags": [
          "Cloudflare R2 Media Management"
        ],
        "summary": "Get Media Access",
        "description": "Cấp URL truy cập ngắn hạn cho file riêng tư sau khi xác thực quyền người dùng.",
        "operationId": "get_media_access_api_media__media_id__access_get",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/{media_id}": {
      "delete": {
        "tags": [
          "Cloudflare R2 Media Management"
        ],
        "summary": "Delete Media",
        "description": "Xóa file khỏi R2 và Supabase.",
        "operationId": "delete_media_api_media__media_id__delete",
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/local-upload": {
      "post": {
        "tags": [
          "Cloudflare R2 Media Management"
        ],
        "summary": "Handle Local Upload",
        "description": "Endpoint dev lưu file cục bộ khi chạy offline.",
        "operationId": "handle_local_upload_api_media_local_upload_post",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viet-phuc-public",
              "title": "Bucket"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_handle_local_upload_api_media_local_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/files/{bucket}/{file_path}": {
      "get": {
        "tags": [
          "Cloudflare R2 Media Management"
        ],
        "summary": "Serve Local File",
        "description": "Endpoint dev phục vụ file tĩnh cục bộ.",
        "operationId": "serve_local_file_api_media_files__bucket___file_path__get",
        "parameters": [
          {
            "name": "bucket",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bucket"
            }
          },
          {
            "name": "file_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccessUrlResponse": {
        "properties": {
          "access_url": {
            "type": "string",
            "title": "Access Url"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_url",
          "expires_in"
        ],
        "title": "AccessUrlResponse"
      },
...
      "Body_handle_local_upload_api_media_local_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_handle_local_upload_api_media_local_upload_post"
      },
...
      "CompareRequest": {
        "properties": {
          "snapshot_a": {
            "$ref": "#/components/schemas/OutfitSnapshot-Input"
          },
          "snapshot_b": {
            "$ref": "#/components/schemas/OutfitSnapshot-Input"
          }
        },
        "type": "object",
        "required": [
          "snapshot_a",
          "snapshot_b"
        ],
        "title": "CompareRequest"
      },
      "CompareResponse": {
        "properties": {
          "diffs": {
            "items": {
              "$ref": "#/components/schemas/SlotDiff"
            },
            "type": "array",
            "title": "Diffs"
          },
          "style_changed": {
            "type": "boolean",
            "title": "Style Changed"
          },
          "occasion_changed": {
            "type": "boolean",
            "title": "Occasion Changed"
          },
          "summary_message": {
            "type": "string",
            "title": "Summary Message"
          }
        },
        "type": "object",
        "required": [
          "diffs",
          "style_changed",
          "occasion_changed",
          "summary_message"
        ],
        "title": "CompareResponse"
      },
      "CompleteUploadRequest": {
        "properties": {
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          }
        },
        "type": "object",
        "title": "CompleteUploadRequest"
      },
...
      "CreateOutfitRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": "Bản phối mới"
          },
          "occasion_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occasion Id"
          },
          "style_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Mode"
          },
          "snapshot": {
            "$ref": "#/components/schemas/OutfitSnapshot-Input"
          },
          "preview_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Image Url"
          }
        },
        "type": "object",
        "required": [
          "snapshot"
        ],
        "title": "CreateOutfitRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaAssetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "public_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Url"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "bucket",
          "object_key",
          "media_type",
          "mime_type",
          "visibility",
          "status",
          "created_at"
        ],
        "title": "MediaAssetResponse"
      },
...
      "OutfitResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "occasion_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occasion Id"
          },
          "style_mode": {
            "type": "string",
            "title": "Style Mode"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "current_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "current_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutfitSnapshot-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "preview_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Image Url"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "style_mode",
          "revision",
          "created_at",
          "updated_at"
        ],
        "title": "OutfitResponse"
      },
...
      "RequestUploadUrlInput": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type",
            "default": "image"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type",
            "default": "image/png"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility",
            "default": "private"
          }
        },
        "type": "object",
        "required": [
          "filename"
        ],
        "title": "RequestUploadUrlInput"
      },
...
      "UpdateOutfitRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "occasion_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occasion Id"
          },
          "style_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Mode"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "snapshot": {
            "$ref": "#/components/schemas/OutfitSnapshot-Input"
          },
          "preview_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Image Url"
          }
        },
        "type": "object",
        "required": [
          "revision",
          "snapshot"
        ],
        "title": "UpdateOutfitRequest"
      },
...
      "UploadUrlResponse": {
        "properties": {
          "media_id": {
            "type": "string",
            "title": "Media Id"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "storage_type": {
            "type": "string",
            "title": "Storage Type"
          }
        },
        "type": "object",
        "required": [
          "media_id",
          "upload_url",
          "method",
          "object_key",
          "bucket",
          "expires_in",
          "storage_type"
        ],
        "title": "UploadUrlResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}